requirements.txt
google-generativeai
pydantic
pytest
requests
PyYAML
flask
dataclasses
python-dotenv
